chore(payments): cleanup of pre-SCA code in payments-server
fixes #5928
