--- packages/fxa-payments-server/src/components/PaymentFormV2/index.scss
@@ -1,97 +1,0 @@
-@import '../../../../fxa-content-server/app/styles/breakpoints';
-
-form.payment {
-  .legal-blurb {
-    clear: both;
-    color: #686869;
-    font-size: 12px;
-    font-weight: 500;
-    line-height: 16px;
-    text-align: center;
-
-    p {
-      margin-bottom: 4px;
-      margin-top: 0;
-    }
-
-    p:last-child {
-      margin-bottom: 0;
-    }
-  }
-
-  hr {
-    margin: 24px 0 0;
-
-    @include respond-to('simpleSmall') {
-      margin-top: 16px;
-    }
-  }
-
-  .input-row {
-    margin: 24px 0 0;
-
-    @include respond-to('simpleSmall') {
-      margin-top: 16px;
-    }
-
-    input {
-      font-size: 16px;
-
-      // overrides firefox's default box shadow for invalid els
-      // we get this from our class system
-      &:invalid {
-        box-shadow: none;
-      }
-
-      &:-moz-submit-invalid {
-        box-shadow: none;
-      }
-
-      &:-moz-ui-invalid {
-        box-shadow: none;
-      }
-    }
-
-    input::placeholder {
-      color: #767676 !important;
-      font-size: 16px;
-      font-weight: 400;
-      height: 45px;
-      line-height: normal;
-    }
-
-    input::-moz-placeholder {
-      line-height: 45px;
-    }
-  }
-
-  button {
-    margin-top: 32px;
-  }
-
-  .lock::before {
-    background-image: url('./images/lock.svg');
-    background-position: 0 2px;
-    background-repeat: no-repeat;
-    content: '\00a0';
-    display: inline-block;
-    height: 18px;
-    margin: 0 8px;
-    width: 14px;
-  }
-}
-
-@include respond-to('simpleSmall') {
-  .StripeElement .InputContainer .InputElement::-moz-placeholder,
-  form.payment .input-row input::-moz-placeholder {
-    line-height: 38px !important;
-  }
-
-  .main-content .payments-card {
-    width: 100%;
-
-    .input-row-group {
-      padding-bottom: 8px;
-    }
-  }
-}
